sdk/ruby: test against multiple SDK versions
We use Appraisal gem to populate a gemfile for each of the ruby sdk
versions we want to test against. Since ruby tries to load the source files
in `ruby/lib` directory every time we do `with-ledgerd bundle exec
appraisal <version> rspec`, we have to delete `ruby/lib` from the
`$LOAD_PATH` except when we want to test against the latest version
(HEAD).

The new test cases are placed in the subdirectories based on their version,
i.e. spec/integration/1.1 or spec/integration/1.2. SDK version 1.2 shall
pass all the test files with the version number that are less and equal to
1.2.

Closes #2219

Author: Howard Chen <howard7708@gmail.com>
Date: Tue Feb 13 10:38:53 2018 -0800
upstream:65a165921917223ae1e01401aa57809569a6c021