sdk/ruby: test last_page on page objects
Test the boundary case of the second-to-last page being full
and the last page being empty.

We need to bump up against the end of the stream to know it's there.

This is the same behavior as the Unix read function
and Go's `io.Reader` interface.

https://godoc.org/io#Reader

Closes #2986

Author: Dan Croak <dan@statusok.com>
Date: Thu Mar 22 12:28:24 2018 -0700
upstream:9c596029f8699e2a088547cd6f3a4acec8f79879