sdk/ruby: refresh session based on TTL
If we ever change the hostname for a ledger,
we don't want client software to need to be restarted
to pick up the new value.

The `/hello` endpoint returns a TTL for this purpose.
When the TTL expires, re-build the HTTP client
used for all API requests in the SDK.

If there is a problem with subsequent `/hello` requests,
it is okay for the SDK to continue using the address it already knows
while it continues trying for a successful `/hello`.
However, if the SDK has a problem making its initial `/hello` request,
it should report an error to the application.

`Session` is the only object that uses `team_name` and `addr`.
Move `/hello` into `Session` to manage this state.

Running hello in a background thread means it runs
concurrently with the application's thread(s), and it needs
to update some instance variables.

Instance variables are not safe for concurrent access,
so we must synchronize all reads and writes to shared
instance variables with a mutex.

If we fail to do this, things will keep working right up
until they break, and then it will be very hard to debug.

Extract a `Hello` object for testability.

Closes #4069

Author: Dan Croak <dan@statusok.com>
Date: Wed Jul 11 15:56:47 2018 -0700
upstream:fde0eaa191fe25fa9ffce17f6e04e3eddc2b3cec